New command accessibility category for admins

+++ apps/communitybot/src/commands/informatics/removewebhook.js
@@ -12,10 +12,12 @@ module.exports = {
         accessibility: "Moderator",
     },
     run: async (client, message, args) => {
-        if (!args[0]) return message.reply("Provide a Webhook ID")
-        if (!args[1]) return message.reply("Provide a Webhook token")
         message.delete()
         message.reply("Message removed to prevent unauthorized webhook access")
+        if (!message.member.hasPermission("MANAGE_WEBHOOKS")) return message.reply("Nice try! You need the `MANAGE_WEBHOOKS` permission!")
+        if (!args[0]) return message.reply("Provide a Webhook ID")
+        if (!args[1]) return message.reply("Provide a Webhook token")
+        
         const webRaw = await fetch(`${apiurl}/informatics/removewebhook`, {
             method: "DELETE",
             body: JSON.stringify({
